requirements_dev.txt
Authomatic==0.0.10
Flask==0.10.1
Flask-Cache==0.13.1
Flask-DebugToolbar==0.9.0
Flask-GAE-Mini-Profiler==0.1.2
Flask-WTF==0.9.0
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
WTForms==2.0.1
Werkzeug==0.9.6
argparse==1.2.1
blinker==1.3
gnureadline==6.3.3
google-api-python-client==1.3.1
httplib2==0.9
ipdb==0.8
ipython==2.3.0
itsdangerous==0.24
oauth2client==1.4.6
pdb==0.1
pyasn1==0.1.7
pyasn1-modules==0.0.5
python-gnupg==0.3.7
rsa==3.1.4
simplejson==3.6.5
six==1.8.0
uritemplate==0.6
wsgiref==0.1.2
